performance by smaller reflection scan. added mac special OSFileStore handling. rename for clarity
