+++ oshi-core/src/main/java/oshi/software/os/unix/freebsd/FreeBsdFileSystem.java
@@ -32,11 +32,11 @@ import java.util.Map;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.software.common.AbstractFileSystem;
-import oshi.software.common.OSFileStoreInterface;
+import oshi.software.common.specialosfs.OSFileStoreInterface;
 import oshi.software.os.OSFileStore;
 import oshi.software.os.linux.LinuxOSFileStore;
 import oshi.util.ExecutingCommand;
-import oshi.util.FileSystemUtil;
+import oshi.util.OSFileStoreUtil;
 import oshi.util.ParseUtil;
 import oshi.util.platform.unix.freebsd.BsdSysctlUtil;
 
@@ -94,7 +94,7 @@ public final class FreeBsdFileSystem extends AbstractFileSystem {
             }
         }
 
-        Map<String, OSFileStoreInterface> specialFileSystems = FileSystemUtil.getSpecialFileSystems();
+        Map<String, OSFileStoreInterface> specialFileSystems = OSFileStoreUtil.getSpecialOSFileStores();
 
         // Get mount table
         for (String fs : ExecutingCommand.runNative("mount -p")) {
