+++ oshi-core/src/main/java/oshi/software/os/unix/solaris/SolarisFileSystem.java
@@ -34,10 +34,10 @@ import com.sun.jna.platform.unix.solaris.LibKstat.Kstat; // NOSONAR
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.software.common.AbstractFileSystem;
-import oshi.software.common.OSFileStoreInterface;
+import oshi.software.common.specialosfs.OSFileStoreInterface;
 import oshi.software.os.OSFileStore;
 import oshi.util.ExecutingCommand;
-import oshi.util.FileSystemUtil;
+import oshi.util.OSFileStoreUtil;
 import oshi.util.ParseUtil;
 import oshi.util.platform.unix.solaris.KstatUtil;
 import oshi.util.platform.unix.solaris.KstatUtil.KstatChain;
@@ -96,7 +96,7 @@ public class SolarisFileSystem extends AbstractFileSystem {
             }
         }
 
-        Map<String, OSFileStoreInterface> specialFileSystems = FileSystemUtil.getSpecialFileSystems();
+        Map<String, OSFileStoreInterface> specialFileSystems = OSFileStoreUtil.getSpecialOSFileStores();
 
         // Get mount table
         for (String fs : ExecutingCommand.runNative("cat /etc/mnttab")) { // NOSONAR squid:S135
