+++ oshi-core/src/main/java/oshi/software/os/windows/WindowsFileSystem.java
@@ -41,7 +41,9 @@ import oshi.driver.windows.perfmon.ProcessInformation.HandleCountProperty;
 import oshi.driver.windows.wmi.Win32LogicalDisk;
 import oshi.driver.windows.wmi.Win32LogicalDisk.LogicalDiskProperty;
 import oshi.software.common.AbstractFileSystem;
+import oshi.software.common.specialosfs.OSFileStoreInterface;
 import oshi.software.os.OSFileStore;
+import oshi.util.OSFileStoreUtil;
 import oshi.util.ParseUtil;
 import oshi.util.platform.windows.WmiUtil;
